fix cpu baseline kernel bug for rectangle pattern

+++ xgminer/include/kernel.h
@@ -16,15 +16,15 @@ public:
     ~Kernel() {}
 
     void rectangle4_baseline_cpu_kernel(int vertices, std::vector<std::set<int>> edgeLists,
-                                        int& total_count, std::vector<int>& embedding);
+                                        long long& total_count, std::vector<int>& embedding);
 
     void rectangle4_baseline_cpu_kernel_with_graphpi_sched(int vertices, std::vector<std::set<int>>& edgeLists,
-                                        int& total_count, std::vector<int>& embedding, 
+                                        long long& total_count, std::vector<int>& embedding, 
                                         std::vector<std::vector<int>>& p_edgeList, 
                                         std::vector< std::pair<int,int>>& restrict_pair);
 
     void rectangle4_baseline_cpu_kernel_with_graphpi_sched_v2(int vertices, std::vector<std::set<int>>& edgeLists,
-                                        int& total_count, std::vector<int>& embedding, 
+                                        long long& total_count, std::vector<int>& embedding, 
                                         std::vector<std::vector<int>>& p_edgeList, 
                                         std::vector< std::pair<int,int>>& restrict_pair);
